CLN: move example code out of script/module
